+++ encoder/enc_amt22.h
@@ -1,5 +1,5 @@
 /*
-	Copyright 2022 Benjamin Vedder	benjamin@vedder.se
+	Copyright 2026 Benjamin Vedder	benjamin@vedder.se
 
 	This file is part of the VESC firmware.
 
@@ -17,22 +17,18 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef ENCODER_CFG_H_
-#define ENCODER_CFG_H_
+#ifndef ENC_AMT22_H_
+#define ENC_AMT22_H_
 
-#include "encoder_datatype.h"
+#include "encoder/encoder_datatype.h"
 
-// Global encoder configurations
-extern ABI_config_t encoder_cfg_ABI;
-extern AD2S1205_config_t encoder_cfg_ad2s1205;
-extern AS5x47U_config_t encoder_cfg_as5x47u;
-extern AS504x_config_t encoder_cfg_as504x;
-extern BISSC_config_t encoder_cfg_bissc;
-extern MT6816_config_t encoder_cfg_mt6816;
-extern ENCSINCOS_config_t encoder_cfg_sincos;
-extern TLE5012_config_t encoder_cfg_tle5012;
-extern TS5700N8501_config_t encoder_cfg_TS5700N8501;
-extern ma782_config_t encoder_cfg_ma782;
-extern AMT22_config_t encoder_cfg_amt22;
+// Functions
+bool enc_amt22_init(AMT22_config_t *AS504x_config);
+void enc_amt22_deinit(AMT22_config_t *cfg);
+void enc_amt22_routine(AMT22_config_t *cfg);
+float enc_amt22_read_angle(AMT22_config_t *cfg);
 
-#endif /* ENCODER_CFG_H_ */
+// Macros
+#define AMT22_LAST_ANGLE(cfg)		((cfg)->state.last_enc_angle)
+
+#endif /* ENC_AMT22_H_ */