+++ encoder/encoder.h
@@ -35,6 +35,7 @@
 #include "enc_ts5700n8501.h"
 #include "enc_pwm.h"
 #include "enc_ma782.h"
+#include "enc_amt22.h"
 
 // Functions
 bool encoder_init(volatile mc_configuration *conf);