+++ hwconf/vesc/minim/hw_minim_core.c
@@ -582,7 +582,7 @@ static THD_FUNCTION(smart_switch_thread, arg) {
 			break;
 
 		case SWITCH_HELD_AFTER_TURN_ON:
-			if (smart_switch_is_pressed()) {
+			if (smart_switch_is_pressed() && conf->shutdown_mode != SHUTDOWN_MODE_ALWAYS_OFF) {
 				switch_state = SWITCH_HELD_AFTER_TURN_ON;
 			} else {
 				switch_state = SWITCH_TURNED_ON;
@@ -591,6 +591,7 @@ static THD_FUNCTION(smart_switch_thread, arg) {
 
 		case SWITCH_TURNED_ON:
 			if (conf->shutdown_mode == SHUTDOWN_MODE_ALWAYS_OFF) {
+				switch_bright = 1.0;
 				if (!smart_switch_is_pressed()) {
 					switch_state = SWITCH_SHUTTING_DOWN;
 				}
@@ -614,9 +615,20 @@ static THD_FUNCTION(smart_switch_thread, arg) {
 
 		case SWITCH_SHUTTING_DOWN:
 			switch_bright = 0;
+			systime_t tStart = chVTGetSystemTimeX();
 			while (smart_switch_is_pressed()) {
 				chThdSleepMilliseconds(10);
+				if (UTILS_AGE_S(tStart) > 10.0) {
+					switch_pressed_ts = chVTGetSystemTimeX();
+					switch_state = SWITCH_TURNED_ON;
+					break;
+				}
 			}
+
+			if (switch_state == SWITCH_TURNED_ON) {
+				break;
+			}
+
 			comm_can_shutdown(255);
 			mc_interface_set_current(0);
 			mc_interface_lock();