+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 * Sin/Cos encoder filter delay compensation.
 * Moved field weakening to fast loop for better stability.
 * New AB encoder sensor mode for encoders without index pulse.
+* Added MA782 encoder support. See https://github.com/vedderb/bldc/pull/886
+* Added AMT22 encoder support.
+* LispBM
+	* Performance improvements and bug fixes
+	* AS5047 encoder pin remapping support
+	* Hall sensor pin remapping support
 
 ### 6.06.6
 #### 2025-11-01